Add deps.edn
It declares dependency on morfologik-fsa, a Java library implementing
native support for DAWGs (aka FSAs) in Jan Daciuk’s formats.

This 12-year-old code used to support reading a different format, one
that I’ve lost track of how to generate. It used clj-bitfields so this
commit will not compile; I’ll replace it with due course.